Update patch set 2
Patch Set 2:

The tests are run with a tycho-p2-director-plugin:2.1.0 because its version is not specified in the testpomless poms (tycho-extras-its/src/test/resources/testpomless/pom.xml etc.).

I could fix the tests by adding tycho-p2-director-plugin in <pluginManagement>. Also I removed the undefined <pluginRepositories> which significantly speeds up the tests.

Should I open a bug for that? Are there other tycho plugins that should have a fixed version? And can we do that somewhere in the test harness instead of touching every test pom?

Patch-set: 2