Update patch set 5
Patch Set 5:

> Maybe MavenProject.setContextValue should *also* be made threadsafe, I'm just not sure if it is intentional. Also as mentioned before I think making threadsafe is often not enough, either way I think its good to access the data in a consistent manner through ReactorProject.

As far as I understand and unless I missed something, it was equally consistent before. So I don't think this change brings a consistency benefit.
Would having the MavenProject.setContextValue(...) operation threadsafe be a good fix here as well? If yes, please first open the enhancement proposal to Maven and link from Tycho code to Maven ticket where is makes sense, so we know we can remove this extra code in the future when Maven works better for Tycho.

Patch-set: 5