Update patch set 5
Patch Set 5:

@Julian:
I think the number of changes could be reduced (and made more clean what really changed) if we add a method
> TychoProjectUtils.getDependencyArtifacts(MavenProject)
that delegates to
> TychoProjectUtils.getDependencyArtifacts(DefaultReactorProject.adapt(mavenProject))

that way callers do not need to mind where the data is actually stored (MavenProject vs ReactorProject) if we want to change this in the future (e.g. MavenProject is made threadsafe).

Patch-set: 5