Update patch set 5
Patch Set 5:

> @Julian:
> I think the number of changes could be reduced (and made more clean what really changed) if we add a method
> > TychoProjectUtils.getDependencyArtifacts(MavenProject)
> that delegates to
> > TychoProjectUtils.getDependencyArtifacts(DefaultReactorProject.adapt(mavenProject))
> 
> that way callers do not need to mind where the data is actually stored (MavenProject vs ReactorProject) if we want to change this in the future (e.g. MavenProject is made threadsafe).

I can split this commit if desired, but that delegation is the wrong approach IMHO.
If tycho adds its own abstraction over MavenProject, that abstraction should be used consistently - not a mix of both.

> As far as I understand and unless I missed something, it was equally consistent before. So I don't think this change brings a consistency benefit.

Data is now consistently written and read via the ReactorProject API.

> Would having the MavenProject.setContextValue(...) operation threadsafe be a good fix here as well? 

Yes, that would solve this issue as well.

Patch-set: 5