Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > @Julian:
> > I think the number of changes could be reduced (and made more clean what really changed) if we add a method
> > > TychoProjectUtils.getDependencyArtifacts(MavenProject)
> > that delegates to
> > > TychoProjectUtils.getDependencyArtifacts(DefaultReactorProject.adapt(mavenProject))
> > 
> > that way callers do not need to mind where the data is actually stored (MavenProject vs ReactorProject) if we want to change this in the future (e.g. MavenProject is made threadsafe).
> 
> I can split this commit if desired, but that delegation is the wrong approach IMHO.

I think splitting is not necessary.

> If tycho adds its own abstraction over MavenProject, that abstraction should be used consistently - not a mix of both.

some code sits in the Maven World and one in an (embedded) OSGi fraework so there are some interactions needed, I think the keypoint is that we store it consistent and make it as easy as possible to work with it.

TychoProjectUtils is a util class to make things easier so I won't mind to much here, its not much difference either to call TychoProjectUtils(...) or TychoProjectUtils(DefaultReactorProject.adapt(mavenProject)) it just clutters the code

> 
> > As far as I understand and unless I missed something, it was equally consistent before. So I don't think this change brings a consistency benefit.
> 
> Data is now consistently written and read via the ReactorProject API.
> 
> > Would having the MavenProject.setContextValue(...) operation threadsafe be a good fix here as well? 
> 
> Yes, that would solve this issue as well.

As mentioned in previous comment I think using the MavenProject.get/setContextValue is not really good design and we should avoid it whenever possible.

Patch-set: 5