Update patch set 5
Patch Set 5:

> Yes that's what I have thought of, I wonder if we really need the ReactorProject variant anyways, it seems TychoProjectUtils is only useful on the "maven side".

Quite the opposite: Most of the callers already have a ReactorProject available. Here's the number of calls that adapt exclusively to use the utility, e.g. 
    TychoProjectUtils.getDependencyArtifacts(DefaultReactorProject.adapt(project))

* getDependencyArtifacts: 2/5 
* getTargetPlatformConfiguration: 5/17
* getTargetPlatform: 2/4
* getTargetPlatformIfAvailable: 1/3
* getExecutionEnvironmentConfiguration: 4/9
* getDependencySeeds: 1/3

Also from an OOP standpoint you could argue that TychoProjectUtils should be replaced with equivalent instance methods on ReactorProject. That's why I believe my change is a step in the right direction.

Patch-set: 5