Update patch set 6
Patch Set 6:

> Moving the utilities is not as easy as I hoped. 
> the relevant types (DependencyArtifacts, TargetPlatformConfiguration, ...) are not available in org.eclipse.tycho.embedder.shared

Too bad. Let's move what's easy to move and comment on the remaining parts that cannot be moved that those should eventually be moved to ReactorProject and the reason currently preventing that from happening. That should be enough for this iteration.
Later, we should consider the merge of org.eclipse.tycho.embedder.core and org.eclipse.tycho.core.shared. Please bring this to tycho-dev mailing-list for discussion, for future releases.

Patch-set: 6