Update patch set 6
Patch Set 6:

That's because of the double-face of tycho, that has hurt me also in the past.
The problem is, that we don't made the maven API available to the OSGi framework running, I could think about ways to accomplish this but I think that's way too much here.

We should fallback to simply store the data in the reactor project and keep the helper for now.

Patch-set: 6