Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

I started with some finer grain review (which you can find inlined) but then realized there is something bigger:
That seems somehow redundant with some of the duties/capabilities of https://www.eclipse.org/tycho/sitedocs/tycho-extras/tycho-p2-extras-plugin/mirror-mojo.html . I would expect that instead of a new mojo, people who need this kind of a customization should be able to do it with this mirror-mojo. So IMO, it would be better to investigate whether just expanding the mirror mojo to provide this content-only rebuild can be done.
If this cannot be done as part of this mirror-mojo, I think this mojo should be moves in the tycho-p2-extras-plugin to somehow highlight it's a more hacky area than what Tycho does by default.
We also need integration test showcase this feature in action, to verify it works, and also to serve as a demo for adopters.

Patch-set: 1
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1