Update patch set 1
Patch Set 1:

(1 comment)

So this just moves what MirrorApplicationServiceImpl.java  already did in its own thing?

So this patch fixes the p2 repository creation we we make it from a maven repo (when the artifacts are not there yet)
This does not fix if the artifacts already did come from a p2 repo (so metadata is already there)?

Patch-set: 1
CC: Gerrit User 16528 <16528@97ee7c02-f12f-4043-b43e-dea463d88b31>