Update patch set 1
Patch Set 1:

(2 comments)

Thank Mickael + Johan for taking a look

> Patch Set 1: Code-Review-1
> 
> (3 comments)
> 
> I started with some finer grain review (which you can find inlined) but then realized there is something bigger:
> That seems somehow redundant with some of the duties/capabilities of https://www.eclipse.org/tycho/sitedocs/tycho-extras/tycho-p2-extras-plugin/mirror-mojo.html .
Mirror Mojo uses MirrorApplication (from P2) while here we use MirrorApplicationService (tycho) what sounds similar and internally uses for some parts MirrorApplication but does a different thing as far as I understand.

> I would expect that instead of a new mojo, people who need this kind of a customization should be able to do it with this mirror-mojo. 
I would be happy to see an example with existing codebase! According to the docs you can do a lot but with the mirror mojo but I can't tell how one would regenerate the metadata with it.

> If this cannot be done as part of this mirror-mojo, I think this mojo should be moves in the tycho-p2-extras-plugin to somehow highlight it's a more hacky area than what Tycho does by default.
The mojo just splits work done in tycho already these days and can't be moved away. Its currently just not possible to use that code

> We also need integration test showcase this feature in action, to verify it works, and also to serve as a demo for adopters.
We could use m2e Bug 570315 about signing as a demo, this clearly shows the usecase Johan has also mentined.

> Patch Set 1:
> 
> (1 comment)
> 
> So this just moves what MirrorApplicationServiceImpl.java  already did in its own thing?
> 
> So this patch fixes the p2 repository creation we we make it from a maven repo (when the artifacts are not there yet)
> This does not fix if the artifacts already did come from a p2 repo (so metadata is already there)?

It dosen't make a difference where the artifacts are from, technically the process is the follwoing:
- tycho creates a repository containing just meta-data (the unis the repository references to be included) plus some java.se.X units
- then it creates a "mirror" that copies everything into "repository" folder
- then it regenerates the artifact repository metadata (that is: file size, hash sum) this step is referenced as neccesary by Bug 357513

Just to summarize, this does not do anything special or new! This (and the previous) patch just splits the code and move the phases a bit to archive the same thing with the same code as before just in a manner that makes it possible to "slip in" with standard maven techniques, either pom-based or with a custom mojo to modify the existing content of the artifact-repository (mtadata is NOT changed!) before packaging the final data.

What one must do then is:
- attach the custom/standard/whatever mojo to the prepare-package-phase
- attatch the rebuild-artifacts-metadata to rebuild metadata (again) this is simimalr to the aproach when generating a source-feature where one needs to call the attached-p2-metadata again to have it recognized the generated source-feature.

Patch-set: 1