Update patch set 1
Patch Set 1:

(2 comments)

> Mirror Mojo uses MirrorApplication (from P2) while here we use MirrorApplicationService (tycho)

First line of MirrorMojo.execute(...) is  `final MirrorApplicationService mirrorService = p2.getService(MirrorApplicationService.class)`.

> I would be happy to see an example with existing codebase! According to the docs you can do a lot but with the mirror mojo but I can't tell how one would regenerate the metadata with it.

Indeed, the goal and customization given seem too different to make it worth factorizing in a single Mojo. A new one -as you suggested- seems better.

> The mojo just splits work done in tycho already these days and can't be moved away. Its currently just not possible to use that code

As mentioned inline, we cannot really split as there are assumption about the results of the mojos; and splitting the mirroring breaks them. Instead of splitting, we need hook for "overriding" some default values.

> We could use m2e Bug 570315 about signing as a demo, this clearly shows the usecase Johan has also mentined.

That would be great, but we cannot really sign as part of the build if we want to make it portable. But replacing the signing step by something that just adds 1 file into the jar or alter the artifact in any way that is not breaking and changes checksums should be easier and more portable.

Patch-set: 1
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>