Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> > Mirror Mojo uses MirrorApplication (from P2) while here we use MirrorApplicationService (tycho)
> 
> First line of MirrorMojo.execute(...) is  `final MirrorApplicationService mirrorService = p2.getService(MirrorApplicationService.class)`.

Yes but MirrorMojo calls 'mirrorStandalone' what neither calls update-metadata nor xyzCompress, so it seems to me more limited and I see no way to call the (private) update methods required here! These actually call the "RecreateRepositoryApplication" so from a design point of view this is probably bad and should be happen in an RecreateRepositoryApplicationService anyways!

> > I would be happy to see an example with existing codebase! According to the docs you can do a lot but with the mirror mojo but I can't tell how one would regenerate the metadata with it.
> 
> Indeed, the goal and customization given seem too different to make it worth factorizing in a single Mojo. A new one -as you suggested- seems better.

okay, so I'll simply create a new mojo and keep the old untouched right?

> > We could use m2e Bug 570315 about signing as a demo, this clearly shows the usecase Johan has also mentined.
> 
> That would be great, but we cannot really sign as part of the build if we want to make it portable. But replacing the signing step by something that just adds 1 file into the jar or alter the artifact in any way that is not breaking and changes checksums should be easier and more portable.

It seems there is already [1] so it seems there is at laest some kind of signing already available as a test.

[1] https://git.eclipse.org/r/c/tycho/org.eclipse.tycho/+/59676/2/tycho-its/src/test/java/org/eclipse/tycho/test/jarsigning/JarSigningTest.java

Patch-set: 1