Update patch set 2
Patch Set 2:

Thanks for your support and for reviewing.

As a side-note (maybe you have already concluded from the other change):
From AbstractMojoTestCase only the lookupConfiguredMojo() method configure the mojo in a way it is initialized with the default values correctly. The other lookupMojo.. methods and even configureMojo() don't to it correctly, as far as I can see it.
But the configureMojoWithDefaultConfiguration() from AbstractTychoMojoTestCase can now be used to configure a mojo properly (with default-values) after its creation.

Patch-set: 2