Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Is this patch still necessary with https://git.eclipse.org/r/c/oomph/org.eclipse.oomph/+/175411 ?

Yes the Oomph change was just used to form an example project for testing/profiling. The issue is not specific or related to Oomph at all - I just used it for profiling instead of making a demo build from scratch.

From the screenshots on the bug can see that without this commit (and parent commit) visitProjects is needlessly called over 231000 times but with this patch only 108 times.

Patch-set: 4