Update patch set 2
Patch Set 2:

Thanks Thomas.
I tried adoption the code in the Gerrit change you linked and it does fix the org.eclipse.equinox.security case. However, it seems to now fail when Import-Package is expected to resolve to export-package from a bundle and its fragment (split package between bundle and fragment). See "tycho-core/src/test/resources/projects/fragment-split-package" for example. Maybe the test is wrong and isn't setting the right expectations? ie in case of 1 package exported by the host and the bundles, are both supposed to be listed by `getPackages` ?

Patch-set: 2