Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> >  Tycho does not really support "offline" directly isn't it?
> 
> It does support offline, with some limitations. The OfflineModeTests do actually cover some of the parts that are supposed to work offline.

It seems to rely on the fact that tycho do aggressive caching of artifacts and if an (undiscovered) failure of creating a repo happens uses this. But now obviously that fails as P2 try to request additional data and the error is propagated up.

I can try to take a look the next days, how should we proceed? Should I checkout your patchset 6 and push changes here, should I open a new gerrit?

Patch-set: 8