Update patch set 14
Patch Set 14:

> Patch Set 13:
> 
> This looks a genuine bug in our code trying to look for site.xml always.

I think the real issue is that P2 does not finds the other files it wants and the falls back to site.xml as a last resort. So it is not a bug but more a fallback that kicks in.

Patch-set: 14
Reviewer: Gerrit User 171685 <171685@97ee7c02-f12f-4043-b43e-dea463d88b31>