Update patch set 4
Patch Set 4:

We are creating a Maven Mojo that uses the tycho EclipseRunMojo but preconfigures many of the parameters, for example which repositories to use, which dependencies to include... 
We currently copy the EclipseRunMojo and Repository classes to add the constructors. Adding setters for the fields injected by Maven would also be working for us, if that were preferred by you.

Patch-set: 4
CC: Gerrit User 304513 <304513@97ee7c02-f12f-4043-b43e-dea463d88b31>