Bug 572427 - Cache ModuleWiring wires to speed-up dependency computation
The DependencyComputer is changed to not be a plexus-Component anymore
in order to avoid invalid cache-content.
As component a DependencyComputer may be shared between instances, but
the wires of a ModuleWiring could change while resolving different
modules. Using a new DependencyComputer for each module ensures
isolation/independence.

This avoids also the need for synchronization (in the widest sense) when
modules are resolved concurrently.

Change-Id: I993440e2ebb944be50954945c6a839e7fe945c7b
Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>