Update patch set 4
Patch Set 4:

OK. As Tycho doesn't really maintain API, I think the way you're trying to consume it is dangerous to you: things can break easily and those breakage be ignored.
Have you tried in your mojo to use 

    @Component
    protected PlexusContainer plexus;

then

   plexus.lookup(...);
   plexus.configure(...);

like it's done in AbstractMojoTestCase#lookupMojo(....) -or even directly using this class- to get an instance of the mojo and configure it?
The benefit is that it allows you code to bind with the typical configuration elements of pom.xml, which Tycho maintains more carefully and in a less breaking way than the code.

Patch-set: 4