Update patch set 2
Patch Set 2: Code-Review+1

I looked at the TODO. I don't think it's critical for this case which is currently marginal/experimental. It's more of a limitation.
Note that transitive deps may not be required for classpath in general since classpath is only required for compilation AFAIK and compilation doesn't require transitive deps as long as the type of those transitive deps don't leak into signatures.
So to me, this is a good change to merge as it.
However, if you have the opportunity to write some unit test for this new part (it shouldn't be too hard to query classpath and check bundle from build.properties/additionalBundles is present), that would be welcome as the part of OsgiBundleProject is a hot-spot of Tycho where many refactorings have and will happen, so it's a good hole for potential regressions.

Patch-set: 2
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1