Update patch set 2
Patch Set 2:

I just saw that in the IDE also the dependency ( org.hamcrest.core ) is put on the classpath. This appear in the "non reactor requirements" and I wonder if in general these (as long as they are bundles) should be placed on the classpath what do you think?

Patch-set: 2