Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> 
> Contrary to what I mentioned in the bugs, the wires can change when resolving one module or the other. I don't think those should be cached between modules. DependencyComputer would better not be a component (which might be shared between instances), but should instead be instantiated by the OsgiBundleProject class directly taking the container state in constructor. With that, we have some guarantee of isolation, and cache could still be used, without need to use ConcurrentHashMap.

Understand. I just uploaded a second patch-set that implements your suggestion.
But passing the container state seems not to be necessary. 

Furthermore I changed DependencyComputer to not be a component anymore (in order to avoid that it is sometimes injected as component and somewhere instantiated directly) and adjusted the three places were it were inject accordingly. Or would this be an API break of Tycho?

Additionally I removed the BundleReader that was never used.

Patch-set: 2