Update patch set 3
Patch Set 3:

I have added a testcase now that will fail for now as the dependent bundle is not there. I have investigated a bit more and found that actually the resolver throws away the resolved bundles.

This has lead me to the question: Do we really need this additional Equinox resolver? I know it was recently refactored with a lot of effort but from debugging this case I have the following impression:

- we collect all dependencies with P2
- we pass them to the resolver, taking effort that we restore all requires/provides
- we start a resolve cycle just to get back the list of dependencies we have passed in before

Especially as we are using the runtime resolving to compute the compile-time classpath (what could be different for different reasons) this seems a bit duplicated.

I wondered if you are open that I try to add an option to bypass the additional equinox resolver? Then it would be possible to make some tests how it behaves, at least this should be a lot faster and would resemble more how the project heaves in the IDE.

Patch-set: 3