Update patch set 2
Patch Set 2:

> I have now enabled the "default" surefire plugin also for this and extended the test-case.

Sounds good to add this possibility; although I'm still a bit worried about the risk of consuming undesired API it brings in PDE and so on, but it's a problem of additionBundles design in PDE, not something we can change in Tycho.
One question I didn't manage to figure out from a quick look at the code is how does the maven-surefire-plugin detect the test classes? Are test sources placed into project.build.testOutputDirectory ?
How does it affect the lifecycle of eclipse-test-plugin? Are the classes still in the usual folders independently of the "test" attribute or are they moved to new folders?

Patch-set: 2
Reviewer: Gerrit User 587 <587@97ee7c02-f12f-4043-b43e-dea463d88b31>