Update patch set 2
Patch Set 2:

> Patch Set 2:
> One question I didn't manage to figure out from a quick look at the code is how does the maven-surefire-plugin detect the test classes? Are test sources placed into project.build.testOutputDirectory ?

correct, the maven-surefire-plugin looks there for tests and as this folder is empty for all other cases it then simply skips execution.

> How does it affect the lifecycle of eclipse-test-plugin? Are the classes still in the usual folders independently of the "test" attribute or are they moved to new folders?

I didn't change this for now so it keeps its "old" semantics, if everything works well we might can discourage and/or remove this packaging type completely.

Patch-set: 2