Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > > * Scope=RUNTIME also includes test output jars
> > 
> > Only if the are included in the build.properties in which case they are part of the final artifact and are then compiled twice (once to target/classes and once to target/test-classes)
> 
> Isn't RUNTIME==ALL actually? Eg invoking RUNTIME scope consider all content from .classpath and build.properties?

currently "runtime" is the old build.properties and only TEST uses the classpath to find things not mentioned in build.properties

> Another question about this confusion is what happens if some folder has the "test" flag but is also listed in `source..`? Is this folder part of the main jar?

This will then be included in the main jar, if it makes sense to configure it like this don't know :-)

> Why compiling twice?

Its only compiled twice if it is mentioned in build.properties *and* marked as test-source (once in compile and once in compile-test phase). This combination does not makes much sense so I don't want to optimize for this case.

> Couldn't the related test folders content be copied to project.build.testOutputDirectory by a dedicated mojo? That would IMO be less intrusive in the project lifecycle; basically nothing would change at all and no distinction would have to be made for tests in the OsgiBundleProject and EclipseProjectImpl.

But then basicly it would violate the life-cycle mapping phases. Maven defines a phase where regular clases are compiled (compile) and one where test-sources are compiled (compile-test), mixing those would make things more complicate (e.g. one needs to determin what has to be copied (inner classes? lamdas?...?)

> > Nope. 'eclipse-test-plugin' is simply not used here. This feature only applies to 'eclipse-plugin' 
> 
> OK, so the current approach is that unless there is an extra compilation step for sources, everything remains strictly identical?

As long as one does not define a "test" source folder everything remains the same.

> If so, I believe the suggestion above of a "prepare-test-classes" mojo that would copy the relevant class to the relevant folder would be useful.

Nope, see above :-) This exactly resembles how a plain maven build would behave.

> I think in general, this patch probably does too much at once to make it easy to understand:
> * if fixes some additionalBundle support
this was done in another gerrit already and is not part of this one

> * it adds parsing for .classpath
> * it adds extra folders to compilation according to .classpath
> * it adds some extra goals to eclipse-plugin lifecycle
> * it adds support for Scope and treating tests folder differently

Al those belongs together, while it is possible to split the parsing I think it makes more sense to keep this together as it shows how things are supposed to work to together.

> I'd appreciate if things can be reduced in smaller chunks, that would facilitate review/understanding, and I think it would also allow to find some simplifications.
I can split it anyways if you think it helps.

Patch-set: 2