Update patch set 4
Patch Set 4:

(3 comments)

> Do you still perceive a noticeable performance improvement with this change?

Yes, I can still measure an performance improvement. It was even a few seconds faster, so it is either inaccurate measurement (but the times were quite stable before) or the change from ConcurrentHashMap to HashMap is a bigger performance improvement than the additional cost of re-computing some wires.

Patch-set: 4