Update patch set 4
Patch Set 4:

I think we have a gap here between JDT<->Tycho here :-\

Currently all source-folders for the same output are compiled in one step. This currently works well because in almost all cases we have only one output folder.

But if there are different output folders this won't work anymore. This is because JDT internally seem to tell the compiler for each source-file where to place this. In Tycho we can only specify one output folder at time. PDE seems to "solve" this by jars.compile.order but this makes this feature even more suspicious to me (platform is using this to build jars as standalone+add them to the classpath as an embedded lib...) and requires it to keep an eye that there are no cross references.

The best solution's seems that JDT compiler would accept one output directory per source directory but this seems not be the case currently.

Patch-set: 4