[JUnit5] Separate provider JUnit57 for jupiter and jupiter+vintage
For plain JUnit 5 tests, JUnit vintage is not desired because it
requires to include JUnit 4 in the classpath one way or the other for
plain JUnit 5 test suite to run (or we get a failure while trying to
load JUnitVintage).
So we remove JUnitVintage from the plain JUnit57 support.

For cases where JUnitVintage is required (ie mixing JUnit 4 and JUnit
5.x), we now use a separate provider which does include JUnitVintage and
its extra requirements.

On the long run, the support for JUnitVintage 5.7 may be the last one
provided.