Fix warning from maven-plugin-plugin
It comes as:
'''
[ERROR]
Maven dependencies of Maven Plugins should be in provided scope.
Please make sure that all your dependencies declared in POM whose group
ID is
org.apache.maven have set '<scope>provided</scope>' as well.
In the future this error will break the build.
'''

so just reading the log makes it look like an error which it isn't
(yet!).

There are couple more staying due to maven-surefire-plugin and
maven-gpg-plugin having this bad practice and thus transitively showing
in Tycho build too.