Updated the description of the new goal tycho-surefire:integration-test
I found the description of the new goal tycho-surefire:integration-test in the RELEASE notes a bit confusing and with several errors.

First of all, the RELEASE notes used to mention tycho-failsafe, which does not exist at all. They were also mentioning MOJOs, but maybe it's better to use the term `goals`?

I also tried to summarize the differences (see the "Summarizing..."), which were not clear to me. That's what I understand from taking a look at the code, but I might be wrong.

Is there an example project showing this new goal, by the way?