Replace Hamcrest usage with plain JUnit
Most of the hamcrest usages are just complicating things e.g.
assertThat(object, is(not(nullValue())) instead of
assertNotNull(object).
In addition to that we have a mix of hamcrest 1.x and 2.x due to junit
4/5 mixture.
Extra points for less dependencies in some projects.
This doesn't remove all hamcrest usages possible but the patch is big
enough already so trying to deliver it now.