Remove tycho.mode=maven
This was helpful for cases where tycho resolution is not needed but very
few people are aware of it and thus not widely used. As custom support
to exclude "clean" (probably the most widely used) has been added the
reasons to have it diminished significantly. If there is need for more
to be excluded it should be done like for "clean" rather than rely on
undocumented property.
Fixes #676