Support JUnit5 @DisplayName
If the test provider is JUnitPlatformProvider then set up the Surefire
reporting for JUnit5 as suggested in the Surefire documentation.

In the maven-surefire-plugin this configuration is expressed in the
pom. This approach does not work for Tycho as we have a separate OSGi
process executing the tests.

Instead we use the provider name to determine the reporting. This
doesn't offer the same flexibility as maven-surefire-plugin but should
be sufficient for initial support.

Fixes https://github.com/eclipse/tycho/issues/198