Inject content of settings.xml into artifact repositories.
For artifact repositories which have been specified in the target
platform definition, Tycho creates a minimal instance, containing only
an id and url, which is then injected back into the Maven model.

In order to support e.g. authenticating against a private repository, it
is necessary to also inject the user credentials, provided via the local
settings.xml, into this instance. Otherwise accessing the referenced
artifacts fails with a 401 - Unauthorized.