Embed m2e lifecycle-mapping-metadata in Tycho plug-ins
The embedded metadata simply advice M2E to ignore executions of each
Tycho Mojo.
In the Eclipse IDE Plug-in development support is PDE's duty and
therefore Tycho Plug-ins have nothing to do.
Because the metadata embedded in a plugin are considered with lowest
priority, a user can simply overrule this decision, if it is desired to
execute a Tycho Plugin within the IDE.

Fixes https://github.com/eclipse/tycho/issues/945