Move target resolution to the maven build phase
Currently everything is computed at the project setup phase, this has
some drawbacks:

1) Even if one executes some goals that do not require the resolving of
these artifacts, the computation is performed
2) If computation fails for one module the whole build fails, so flags
like -fae are not really usable and implementing such support directly
is very complex
3) If running a parallel build, the build can't start unless we have
computed all modules
4) if other projects like to contribute items this currently only
possible with the pom=consider mode, as we only allow incomplete
resolving there

This moves the computation of the preliminary target platform into the
project execution phase allowing to relax these steps. This will also
enable us to remove the need for incomplete resoloution option for mixed
builds.
Pass the session as a parameter as with work-stealing it does not work
well to use ThreadLocals!
Ignore requirements that already are satisfied
Make the session required and cone it for concurrency!