Use p2inf in DependencyComputerTest.testFragments
The DependencyComputerTest.testFragments simulates a build that includes
swt+jface but current SWT build uses a p2.inf to require additional
units. This currently makes this test behave a little bit different than
a actual swt build would behave. Because of this, this adds a p2.inf
file that is similar to what swt-build contains.