Use a special hinted COmponent/Requirement for the EquinoxServiceFactory
Currently the plugins in Tycho rely on the default
EquinoxServiceFactory, because of this if we want to switch that many
code needs to be adjusted and it is unlikely that it will work to mix
here. Because of this, there is now a special hinted component that
simply delegates to a EquinoxServiceFactory suitable for use with tycho.