Support arbitrary aggregator directories without extra setup
Removes the need to specify a Set of allowed aggregator-directory names
and simply scans all modules listed (transitively) in the root pom.xml
for directly nested modules, if the directory is not any other kind of
project (and has not pom.xml).

Update to the latest version of io.takari.polyglot:polyglot-common which
reduces the calls to 'Mapping.locatePom()' to the minimum.
Polyglot-maven already only checks those directories listed as modules.
But before it located a pom for all available mappings. Now it starts
with the highest priority and tires those with lower once until a
Mapping.locatePom() returns a result.

To minimize the number of file-system interactions the following
distinct order of Tycho-Mappings is introduced:
- Bundle
- Feature
- pom.xml (provided by polyglot-maven)
- Repository (means Product or p2-repository)
- Target-Platform
- Aggregator

It is important that the Tycho-Feature-Mapping has a higher priority
than the Pom-Mapping. Otherwise it can happen that the Pom-Reader is
used to load the model from the feature.xml file and to populate the
polyglot model, which would then be flawed.

The remaining relations imposed by this order reflect the likelihood of
a match (usually more occurring project types get a higher priority) and
the cost of a match attempt (more expensive mappings get a lower
priority), which results in the aggregator-mapping having the lowest
priority.