Clean-up and improve documentation of baseline comparing Mojos
The 'ignoredPatterns' parameter of the 'p2-metadata'
and 'compare-version-with-baselines' goal is now documented since it
became a mojo parameter.

The property 'tycho.debug.artifactcomparator' is also converted to a
parameter of the tycho-p2-plugin:p2-metadata goal so it can be set in
the pom.

Additionally slightly improve performance of nested Zip-comparision and
of the check if a file is ignored.