Add a surrogate implementation of UI dependencies
There is an optional import for org.eclipse.ui.workbench to support UI
test. If the import can not be resolved, there is no fallback in two
scenarios:

1) we only want compile but do not pull in half of platform
2) the bundle can not be resolved

for those cases the minimal necessary API is copied into the bundle that
will be shaded by the imports if available in the running system.