Avoid BND warning about missing developer id
BND complains: Cannot consider developer in line 'foo' of file
'bar/pom.xml' for bundle header 'Bundle-Developers' as it does not
contain the mandatory id.

To avoid this developers must either have an ID or be removed. Maven
itself considers the list of developers the list of people to be
contacted for the project. Since the currently listed developers are no
longer active, the error can be avoided best (and with little future
maintenance effort) by having an organizational entry as developer,
pointing the Tycho project.